+++ reqcoco-api/src/main/java/net/paissad/tools/reqcoco/api/report/ReqReportConfig.java
@@ -1,22 +1,28 @@
 package net.paissad.tools.reqcoco.api.report;
 
+import java.util.Collection;
+
 import com.google.gson.annotations.Expose;
 
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 public class ReqReportConfig {
 
     @Expose
-    private String title                    = "Requirements Coverage";
+    private String             title                    = "Requirements Coverage";
 
     @Expose
-    private String codeCoverageDiagramName  = "Source Code Coverage";
+    private String             codeCoverageDiagramName  = "Source Code Coverage";
 
     @Expose
-    private String testsCoverageDiagramName = "Tests Code Coverage";
+    private String             testsCoverageDiagramName = "Tests Code Coverage";
 
     @Expose
-    private String requirementsTableLegend  = "Table of requirements";
+    private String             requirementsTableLegend  = "Table of requirements";
 
+    @Expose
+    @Setter
+    private Collection<String> filteredVersions;
 }